+++ src/main/java/utility/StringHelper.java
@@ -1,4 +1,4 @@
-package helper;
+package utility;
 
 /**
  * @author han
