+++ src/test/java/utility/NumberHelperTest.java
@@ -1,6 +1,5 @@
-package Helper;
+package utility;
 
-import helper.NumberHelper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
